requirements.txt
fastapi
uvicorn
websockets
openai
sqlalchemy
pyodbc
termcolor
pandas
azure-data-tables
azure-storage-blob
azure-identity
requests
psycopg2
Jinja2
python-multipart
python-dotenv
python-jose
python-decouple
bcrypt
asyncpg
databases
itsdangerous
pydantic[email]
sqlfluff
numpy
aiofiles
azure-search-documents
aiohttp
plotly
cachetools
PyMuPDF
pymupdf4llm
pypdf==3.17.4
camelot-py
Pillow
sqlparse==0.5.3
qdrant-client
nltk
llama-index
spacy==3.8.4
keyring[full]
keyrings.alt
azure-ai-inference
pytesseract
reportlab
kaleido

# MCP Infrastructure
httpx>=0.27.0
sse-starlette>=2.0.0

# Pydantic AI Framework
pydantic-ai>=0.0.14
pydantic-ai-slim>=0.0.14